requirements.txt
django>=5.0,<6.0
djangorestframework
django-environ
psycopg2-binary
gunicorn
pytest-django
pytest
black
isort
ruff
djangorestframework-simplejwt
drf-nested-routers
django-cors-headers
Pillow
